% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1.R
\name{d1.grid}
\alias{d1.grid}
\title{Calculate grid of approximating points}
\usage{
d1.grid(vRange, iPts, stMethod = "Chebychev")
}
\arguments{
\item{vRange}{the vector of the range over which the nodes are computed}

\item{iPts}{number of colocations points}

\item{stMethod}{Either \code{Chebychev} or \code{Uniform}}
}
\description{
Computes either a Uniform or Chebychev grid of collocation nodes
}

