% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1.R
\name{d1.normalize}
\alias{d1.normalize}
\title{Normalizes x in [a,b] to [-1,1]}
\usage{
d1.normalize(x, range)
}
\arguments{
\item{x}{A number in range}

\item{range}{The range}
}
\value{
2(x-a)/(b-a)-1
}
\description{
Normalizes x in [a,b] to [-1,1].
}

