\name{d1.grid}
\alias{d1.grid}
\title{Calculate grid of approximating points}
\usage{
d1.grid(vRange, iPts, stMethod = "Chebychev")
}
\arguments{
  \item{vRange}{the vector of the range over which the
  nodes are computed}

  \item{iPts}{number of colocations points}

  \item{Either}{\code{Chebychev} or \code{Uniform}}
}
\description{
Computes either a Uniform or Chebychev grid of collocation
nodes
}

